requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
python-multipart==0.0.18
openai==1.58.1
python-docx==1.1.2
pypdf==5.1.0
python-dotenv==1.0.1
httpx==0.28.1
google-api-python-client==2.156.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
beautifulsoup4==4.12.3
Jinja2==3.1.4
aiofiles==24.1.0
